Mark Goodrich closed PIMS-707 (Need to add a javascript "confirm" dialog when removing shipment items from a container in the shipment packing stage) saying: Done
Mark Goodrich closed PIMS-685 (Do not group list of shipments by status) saying:
Changed on both list shipping and list receiving page.

Mark Goodrich closed PIMS-686 (Allow list of shipments to be filtered by type) saying:
Added to list shipping and list receiving pages.

Mark Goodrich closed PIMS-687 (Allow list of shipments to be filtered by status) saying:
Added to both list shipping and list receiving pages.

Mark Goodrich closed PIMS-688 (Allow list of shipments to be filtered by destination) saying:
Added to both list shipping and list receiving pages.

Mark Goodrich closed PIMS-689 (Allow list of shipments to be filtered by date of status (using date range)) saying:
Added to both list shipping and list receiving pages

Mark Goodrich closed PIMS-702 (Modify display of "There are no shipments matching your conditions" on "List all shipping" page) saying:
Moved this message to to below the filter, and removed the error formatting. Changed table header so that it doesn't appear if no entries.

Mark Goodrich closed PIMS-717 (Remove ability to delete events on Show Details page) saying:
Just commented this out for now... since "deleting" a shipped or received event doesn't undo the underlying transaction, we don't want users to be able to delete events. They can, however, still change the date of shipping and receiving events.

Mark Goodrich closed PIMS-718 (Remove "package" section from Show Details section) saying:
Commented this out


git-svn-id: https://dev.pih-emr.org/svn/repository/projects/warehouse/trunk@9852 f0ac6d2b-dfc5-4988-9158-fc266d298d87

+++ grails-app/views/createShipmentWorkflow/createShipment/enterContainerDetails.gsp
@@ -98,7 +98,7 @@
 											${itemInstance?.recipient?.name}
 										</td>
 										<td style="text-align: left;">		
-											<g:link action="createShipment" event="deleteItem" params="['item.id':itemInstance?.id]">	
+											<g:link action="createShipment" event="deleteItem" params="['item.id':itemInstance?.id]" onclick="return confirm('Are you sure you want to delete this item?')">	
 											<img src="${createLinkTo(dir:'images/icons/silk',file:'delete.png')}" alt="remove item" style="vertical-align: middle"/>
 											</g:link>	
 										</td>
@@ -127,7 +127,7 @@
 											</g:link> 		
 										</td>
 										<td nowrap="nowrap">
-											<g:link action="createShipment" event="deleteBox" params="['box.id':boxInstance?.id]">
+											<g:link action="createShipment" event="deleteBox" params="['box.id':boxInstance?.id]" onclick="return confirm('Are you sure you want to delete this box?')">
 												<img src="${createLinkTo(dir:'images/icons/silk',file:'package_delete.png')}" alt="Add an item" style="vertical-align: middle"/>
 											</g:link>				
 										</td>
@@ -153,7 +153,7 @@
 												${itemInstance?.recipient?.name}
 											</td>
 											<td style="text-align: left;" nowrap="true">		
-												<g:link action="createShipment" event="deleteItem" params="['item.id':itemInstance?.id]">
+												<g:link action="createShipment" event="deleteItem" params="['item.id':itemInstance?.id]" onclick="return confirm('Are you sure you want to delete this item?')">
 													<img src="${createLinkTo(dir:'images/icons/silk',file:'delete.png')}" alt="remove item" style="vertical-align: middle"/>
 												</g:link>	
 											</td>
